Handled odd minibatch divisility
Since our network takes in a fixed size minibatch, and our training data
doesn't cleanly fit that, we need to pad the last batch with repeat
data. This is likely going to result in some samples being weighted more
heavily than others during training, but they are randomly sampled and
it is a simple solution for now.